+++ cupyx/scipy/stats/__init__.py
@@ -1,1 +1,1 @@
-from cupyx.scipy.stats.distributions import entropy  # NOQA
+from cupyx.scipy.stats._distributions import entropy  # NOQA
